refactor: Build only ESM, correct TS types
