--- compat/jsx-dev-runtime.mjs
@@ -1,3 +1,0 @@
-import 'preact/compat';
-
-export * from 'preact/jsx-runtime';
