+++ compat/package.json
@@ -1,11 +1,9 @@
 {
   "name": "preact-compat",
-  "amdName": "preactCompat",
   "private": true,
   "description": "A React compatibility layer for Preact",
+  "type": "module",
   "main": "dist/compat.js",
-  "module": "dist/compat.mjs",
-  "umd:main": "dist/compat.umd.js",
   "source": "src/index.js",
   "types": "src/index.d.ts",
   "license": "MIT",
