+++ compat/scheduler.js
@@ -1,15 +1,23 @@
-// see scheduler.mjs
+/* eslint-disable */
 
-function unstable_runWithPriority(priority, callback) {
+// This file includes experimental React APIs exported from the "scheduler"
+// npm package. Despite being explicitely marked as unstable some libraries
+// already make use of them. This file is not a full replacement for the
+// scheduler package, but includes the necessary shims to make those libraries
+// work with Preact.
+
+export var unstable_ImmediatePriority = 1;
+export var unstable_UserBlockingPriority = 2;
+export var unstable_NormalPriority = 3;
+export var unstable_LowPriority = 4;
+export var unstable_IdlePriority = 5;
+
+/**
+ * @param {number} priority
+ * @param {() => void} callback
+ */
+export function unstable_runWithPriority(priority, callback) {
 	return callback();
 }
 
-module.exports = {
-	unstable_ImmediatePriority: 1,
-	unstable_UserBlockingPriority: 2,
-	unstable_NormalPriority: 3,
-	unstable_LowPriority: 4,
-	unstable_IdlePriority: 5,
-	unstable_runWithPriority,
-	unstable_now: performance.now.bind(performance)
-};
+export var unstable_now = performance.now.bind(performance);
