--- compat/test-utils.mjs
@@ -1,1 +1,0 @@
-export * from 'preact/test-utils';
